2_data_ingestion/requirements.txt
prefect
prefect-sqlalchemy
prefect-gcp[cloud_storage]
protobuf
pyarrow
pandas-gbq
psycopg2-binary
sqlalchemy
pgcli
ruamel-yaml
pyqtwebengine==5.12
pyqt5==5.12
psycopg_binary